test(modal): add failing test
